pyproject.toml
[project]
name = "bin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.7",
    "discord-py>=2.7.1",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-groq>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "markdown>=3.10.2",
    "numpy>=2.4.4",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.10.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.0",
    "unstructured>=0.22.18",
]
